requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
nltk>=3.8.0
rouge-score>=0.1.2
pycocoevalcap>=1.2
